// src/types/enums.ts
export enum CHEST_SHAPE {
  NARROW_CHEST = "narrow_chest",
  WIDE_CHEST = "wide_chest",
  MEDIUM_CHEST = "medium_chest",
}

export enum ABDOMEN_SHAPE {
  FLAT_ABDOMEN = "flat_abdomen",
  BULGING_ABDOMEN = "bulging_abdomen",
  MEDIUM_ABDOMEN = "medium_abdomen",
}

export enum BODY_SHAPE {
  APPLE = "apple",
  PEAR = "pear",
  RECTANGLE = "rectangle",
  HOURGLASS = "hourglass",
  INVERTED_TRIANGLE = "inverted_triangle",
}